# swagger-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Home Tuitions API",
    "version": "1.0.0",
    "description": "Enterprise-grade REST API for the Home Tuitions marketplace connecting students/parents with verified expert tutors.\n\n### Authentication\nAll protected endpoints require a valid JWT Bearer Token in the `Authorization` header:\n`Authorization: Bearer <your_jwt_token>`\n\nRate limiting headers and device identification (`X-Device-Id`) are supported across sensitive authentication routes.",
    "contact": {
      "name": "Home Tuitions Engineering Team",
      "url": "https://github.com/bharath042105/Home-tuitions"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.hometuitions.example.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/leads": {
      "post": {
        "tags": ["Leads"],
        "summary": "Submit a tuition inquiry / lead",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateLeadRequest" } }
          }
        },
        "responses": {
          "201": { "description": "Lead created successfully" }
        }
      }
    },
    "/api/v1/leads/upload-url": {
      "post": {
        "tags": ["Leads"],
        "summary": "Request S3/R2 presigned upload URL for tutor registration documents",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["filename", "contentType"],
                "properties": {
                  "filename": { "type": "string", "example": "degree_certificate.pdf" },
                  "contentType": { "type": "string", "example": "application/pdf" },
                  "documentType": { "type": "string", "enum": ["photo", "aadhaar", "degree", "resume"], "example": "degree" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned upload URL and persistent public download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uploadUrl": { "type": "string" },
                    "key": { "type": "string" },
                    "publicUrl": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/tutor-applications": {
      "post": {
        "tags": ["Leads"],
        "summary": "Submit tutor onboarding application with credentials and document URLs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SubmitTutorApplicationRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TutorApplicationResponse" } } }
          }
        }
      }
    },
    "/api/v1/admin/leads/tutor-applications": {
      "get": {
        "tags": ["Admin - Leads"],
        "summary": "List paginated tutor applications with document verification links",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of tutor applications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": { "type": "array", "items": { "$ref": "#/components/schemas/TutorApplicationResponse" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token."
      }
    },
    "schemas": {
...
      "CreateLeadRequest": {
        "type": "object",
        "required": ["subject", "gradeLevel", "address", "preferredSchedule"],
        "properties": {
          "subject": { "type": "string", "example": "Grade 10 ICSE Physics" },
          "gradeLevel": { "type": "string", "example": "Grade 10" },
          "address": { "type": "string", "example": "HSR Layout, Bangalore" },
          "preferredSchedule": { "type": "string", "example": "Weekday Evenings" },
          "budgetMax": { "type": "number", "example": 1000.0 }
        }
      },
...
    },
...
  }
}